Rip out outdated lite and lite_impl modules from tendermint crate
